requirements.txt
django==5.1.7
requests==2.31.0
pandas==2.1.4
numpy==1.26.4
scikit-learn==1.4.0
pytz==2024.1
supabase==2.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
gunicorn==21.2.0
whitenoise==6.7.0
dj-database-url==2.2.0